Update player input handler
