+++ com.unity.cinemachine/Samples~/Input system.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0c786788772714329a368618852f2cda
+guid: 69111a167d68841bca428a7a0e26c953
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
